# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildTrack Construction Expense Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        budget:
          type: number
        laborBudget:
          type: number
        materialBudget:
          type: number
        estimatedRevenue:
          type: number
        startDate:
          type: string
        status:
          type: string
          enum: [active, completed, on_hold]
        totalExpenses:
          type: number
        laborSpent:
          type: number
        materialSpent:
          type: number
        remainingBudget:
          type: number
        profit:
          type: number
        profitMargin:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - budget
        - laborBudget
        - materialBudget
        - estimatedRevenue
        - startDate
        - status
        - totalExpenses
        - laborSpent
        - materialSpent
        - remainingBudget
        - profit
        - profitMargin
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        budget:
          type: number
        laborBudget:
          type: number
        materialBudget:
          type: number
        estimatedRevenue:
          type: number
        startDate:
          type: string
        status:
          type: string
          enum: [active, completed, on_hold]
      required:
        - name
        - budget
        - startDate
        - status
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        budget:
          type: number
        laborBudget:
          type: number
        materialBudget:
          type: number
        estimatedRevenue:
          type: number
        startDate:
          type: string
        status:
          type: string
          enum: [active, completed, on_hold]
...