# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildTrack Construction Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  budget?: number;
  laborBudget?: number;
  materialBudget?: number;
  estimatedRevenue?: number;
  startDate?: string;
  status?: UpdateProjectBodyStatus;
}